General deobfuscation
Decided to make the code at least barely readable
by removing a lot of weird, unnecessary stuff.
That resulted in generally better documentation
and a complete overhaul of the CLI. The program now
accepts multiple source files, but it has less flags,
because some turned out to be useless.

In the meantime, I also changed the name from "huffc"
to "chuff", because it's a better name and I find it
funny.